requirements.txt
langgraph
langchain
openai
python-dotenv
langchain-openai
requests
numpy
pandas
streamlit